requirements.txt
streamlit>=1.40
tensorflow>=2.16
scikit-learn>=1.5
pandas>=2.2
numpy>=1.26
pillow>=10.4
matplotlib>=3.9
